requirements.txt
python-telegram-bot>=20.0
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
asyncpg>=0.29.0
fastapi>=0.110.0
uvicorn>=0.29.0
mangum>=0.17.0
httpx>=0.27.0
